# Conflicts:
#	src/main/java/com/nctc2017/configuration/ApplicationConfig.java
#	src/main/java/com/nctc2017/constants/DatabaseObject.java
#	src/main/java/com/nctc2017/dao/impl/ShipDaoImpl.java
#	src/main/java/com/nctc2017/dao/utils/QueryBuilder.java

+++ src/main/java/com/nctc2017/dao/ShipDao.java
@@ -40,7 +40,7 @@ public interface ShipDao {
 
     List<ShipTemplate> findAllShipTemplates();
 
-    List<Ship> findAllShips(List<Integer> shipsId);
+    List<Ship> findAllShips(List<BigInteger> shipsId);
 
     boolean setMastOnShip(BigInteger mastId, BigInteger shipId);
